transfer keydown listener to main.js
